App background and svg styles
