--- frontend/src/components/app/App.js
@@ -1,24 +1,0 @@
-import './App.css';
-import LandingPage from '../landing/LandingPage';
-import LoginForm from '../auth/LoginForm'
-import SignUpForm from '../user/SignUpForm'
-import React, { useState } from 'react';
-import Feed from '../feed/Feed'
-import {
-  useNavigate,
-  Routes,
-  Route,
-} from "react-router-dom";
-
-const App = () => {
-    return (
-        <Routes>
-          <Route path='/'  element={<LandingPage navigate={ useNavigate() }/>}/>
-          <Route path='/posts'  element={<Feed navigate={ useNavigate() }/>}/>
-          <Route path='/login'  element={<LoginForm  navigate={ useNavigate() }/>}/>
-          <Route path='/signup' element={<SignUpForm navigate={ useNavigate() }/>}/>
-        </Routes>
-    );
-}
-
-export default App;
