+++ frontend/src/components/button/Button.css
@@ -1,6 +1,6 @@
 .button {
-  width: 168px;
-  height: 54px;
+  width: 12rem;
+  height: 4rem;
   border-radius: 8px;
   background: #3360ff;
   box-shadow: 0px 4px 4px rgba(51, 96, 255, 0.65);
@@ -10,4 +10,11 @@
   font-size: 2rem;
   font-weight: bold;
   cursor: pointer;
+}
+
+.button:hover {
+  width: 11.75rem;
+  height: 3.9rem;
+  box-shadow: 0px 5px 5px rgba(51, 96, 255, 0.65);
+  font-size: 1.95rem;
 }
