+++ frontend/src/index.css
@@ -7,7 +7,6 @@ body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   color: #373737;
-  /* background: linear-gradient(rgba(51, 96, 255, 0.26) 52%, #fff 0%); */
 }
 
 code {
